// src/settings/types/banner.enums.ts
export enum BannerPage {
  HOME = 'home',
  CATEGORY_SIDEBAR = 'category_sidebar',
}

export enum BannerSection {
  MAIN_SLIDER = 'main_slider',
  THREE_IMAGES = 'three_images',
  NARROW_BANNER = 'narrow_banner',
  VIDEO = 'video',
  TWO_IMAGES = 'two_images',
  SINGLE_BANNER = 'single_banner',
  SIDEBAR = 'sidebar',
}